requirements.txt
# requirements.txt
# Dependencias de producción

# Web scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# MCP SDK
mcp==1.21.2

# Configuración
python-dotenv==1.0.1
pydantic==2.12.4
pydantic-settings==2.12.0

# Logging mejorado
colorlog==6.8.2

# WhatsApp (Twilio)
twilio==9.0.4

# Utilidades
python-dateutil==2.8.2
pytz==2024.1

# Producción - WSGI Server
gunicorn==21.2.0